Remove unnecessary Endpoint superclass
